Added tests for DataFrames with tenant specific connections and refactored PhoenixSparkIT tests
